fix unneccesary package dependencies
