+++ vuu-ui/packages/vuu-data-react/package.json
@@ -10,18 +10,21 @@
     "type-defs": "node ../../scripts/build-type-defs.mjs"
   },
   "devDependencies": {
-    "@finos/vuu-table-types": "0.0.26",
+    "@finos/vuu-data-types": "0.0.26",
     "@finos/vuu-filter-types": "0.0.26",
     "@finos/vuu-popups": "0.0.26",
-    "@finos/vuu-protocol-types": "0.0.26"
+    "@finos/vuu-protocol-types": "0.0.26",
+    "@finos/vuu-table-types": "0.0.26"
   },
   "dependencies": {
     "@finos/vuu-data-remote": "0.0.26",
-    "@finos/vuu-data-types": "0.0.26",
     "@finos/vuu-filter-parser": "0.0.26",
+    "@finos/vuu-popups": "0.0.26",
+    "@finos/vuu-shell": "0.0.26",
     "@finos/vuu-utils": "0.0.26"
   },
   "peerDependencies": {
-    "react": ">=17.0.2"
+    "react": ">=17.0.2",
+    "react-dom": ">=17.0.2"
   }
 }
