+++ vuu-ui/packages/vuu-table/src/cell-renderers/checkbox-cell/CheckboxCell.tsx
@@ -2,8 +2,11 @@ import React, { memo, useCallback } from "react";
 import { TableCellRendererProps } from "@finos/vuu-table-types";
 import { CheckboxIcon, WarnCommit } from "@finos/vuu-ui-controls";
 import { Checkbox } from "@salt-ds/core";
-import { dataAndColumnUnchanged } from "../cell-utils";
-import { dispatchCustomEvent, registerComponent } from "@finos/vuu-utils";
+import {
+  dataAndColumnUnchanged,
+  dispatchCustomEvent,
+  registerComponent,
+} from "@finos/vuu-utils";
 
 export const CheckboxCell: React.FC<TableCellRendererProps> = memo(
   ({ column, columnMap, onCommit = WarnCommit, row }) => {
