+++ vuu-ui/packages/vuu-table/src/cell-renderers/index.ts
@@ -1,4 +1,3 @@
 export * from "./checkbox-cell";
 export * from "./input-cell";
 export * from "./toggle-cell";
-export * from "./cell-utils";
